+++ snakemake/checkpoints.py
@@ -3,12 +3,15 @@ from snakemake.io import checkpoint_target
 
 
 class Checkpoints:
-    """ A namespace for checkpoints so that they can be accessed via dot notation. """
+    """A namespace for checkpoints so that they can be accessed via dot notation."""
 
     def __init__(self):
         self.future_output = None
 
-    def register(self, rule):
+    def register(self, rule, fallback_name=None):
+        checkpoint = Checkpoint(rule, self)
+        if fallback_name:
+            setattr(self, fallback_name, checkpoint)
         setattr(self, rule.name, Checkpoint(rule, self))
 
 
@@ -27,12 +30,16 @@ class Checkpoint:
             )
 
         output, _ = self.rule.expand_output(wildcards)
-        if self.checkpoints.future_output is None or any(
-            (not f.exists or f in self.checkpoints.future_output) for f in output
-        ):
-            raise IncompleteCheckpointException(self.rule, checkpoint_target(output[0]))
+        if self.checkpoints.future_output is not None:
+            for iofile in output:
+                if iofile in self.checkpoints.future_output:
+                    break
+                if not iofile.exists and not iofile.is_temp:
+                    break
+            else:
+                return CheckpointJob(self.rule, output)
 
-        return CheckpointJob(self.rule, output)
+        raise IncompleteCheckpointException(self.rule, checkpoint_target(output[0]))
 
 
 class CheckpointJob: