+++ snakemake/utils.py
@@ -19,7 +19,7 @@ from urllib.parse import urljoin
 
 from snakemake.io import regex, Namedlist, Wildcards, _load_configfile
 from snakemake.logging import logger
-from snakemake.common import ON_WINDOWS
+from snakemake.common import ON_WINDOWS, is_local_file, smart_join
 from snakemake.exceptions import WorkflowError
 import snakemake
 
@@ -38,8 +38,11 @@ def validate(data, schema, set_default=True):
             https://python-jsonschema.readthedocs.io/en/latest/faq/ for more
             information
     """
-    # skip if a corresponding modifier has been defined
-    if "workflow" in globals() and globals()["workflow"].modifier.skip_validation:
+    frame = inspect.currentframe().f_back
+    workflow = frame.f_globals.get("workflow")
+
+    if workflow and workflow.modifier.skip_validation:
+        # skip if a corresponding modifier has been defined
         return
 
     try:
@@ -51,20 +54,29 @@ def validate(data, schema, set_default=True):
             "in order to use the validate directive."
         )
 
-    if not os.path.isabs(schema):
+    schemafile = schema
+
+    if not os.path.isabs(schemafile):
         frame = inspect.currentframe().f_back
         # if workflow object is not available this has not been started from a workflow
-        if "workflow" in frame.f_globals:
-            workflow = frame.f_globals["workflow"]
-            schema = os.path.join(workflow.current_basedir, schema)
+        if workflow:
+            schemafile = smart_join(workflow.current_basedir, schemafile)
 
-    schemafile = schema
-    schema = _load_configfile(schema, filetype="Schema")
-    resolver = RefResolver(
-        urljoin("file:", schemafile),
-        schema,
-        handlers={"file": lambda uri: _load_configfile(re.sub("^file://", "", uri))},
-    )
+    source = workflow.sourcecache.open(schemafile) if workflow else schemafile
+    schema = _load_configfile(source, filetype="Schema")
+    if is_local_file(schemafile):
+        resolver = RefResolver(
+            urljoin("file:", schemafile),
+            schema,
+            handlers={
+                "file": lambda uri: _load_configfile(re.sub("^file://", "", uri))
+            },
+        )
+    else:
+        resolver = RefResolver(
+            schemafile,
+            schema,
+        )
 
     # Taken from https://python-jsonschema.readthedocs.io/en/latest/faq/
     def extend_with_default(validator_class):
@@ -590,6 +602,9 @@ class Paramspace:
         | ``Paramspace(df, filename_params=["column3", "column2"])`` ->
         | column1~{value1}/column4~{value4}/column3~{value3}_column2~{value2}
 
+        If ``filename_params="*"``, all columns of the dataframe are encoded into
+        the filename instead of parent directories.
+
       - ``param_sep`` takes a string which is used to join the column name and
         column value in the genrated paths (Default: '~'). Example:
 
@@ -606,6 +621,9 @@ class Paramspace:
             self.pattern = "/".join([r"{}"] * len(self.dataframe.columns))
             self.ordered_columns = self.dataframe.columns
         else:
+            if isinstance(filename_params, str) and filename_params == "*":
+                filename_params = dataframe.columns
+
             if any((param not in dataframe.columns for param in filename_params)):
                 raise KeyError(
                     "One or more entries of filename_params are not valid coulumn names for the param file."
@@ -658,8 +676,16 @@ class Paramspace:
         """Obtain instance (dataframe row) with the given wildcard values."""
         import pandas as pd
 
+        def convert_value_dtype(name, value):
+            if self.dataframe.dtypes[name] == bool and value == "False":
+                # handle problematic case when boolean False is returned as
+                # boolean True because the string "False" is misinterpreted
+                return False
+            else:
+                return pd.Series([value]).astype(self.dataframe.dtypes[name])[0]
+
         return {
-            name: pd.Series([value]).astype(self.dataframe.dtypes[name])
+            name: convert_value_dtype(name, value)
             for name, value in wildcards.items()
             if name in self.ordered_columns
         }