+++ CHANGELOG.rst
@@ -1,3 +1,40 @@
+[6.3.0] - 2021-04-29
+====================
+- Changed behavior of ``workflow.snakefile`` to always point to the current file instead of the main Snakefile (also in case of includes and modules) (@johanneskoester).
+- Fixed a typo in an error message (@nikostr).
+
+[6.2.0] - 2021-04-22
+====================
+- Support for integration of foreign workflow management systems by introducing a ``handover`` directive that passes on all resources to a particular rule (which can then invoke another workflow management system). See the docs ("Integrating foreign workflow management systems") (@johanneskoester).
+- Behavior improvement for temp handling of checkpoint rules (@epruesse).
+- Several improvements in the docs (@johanneskoester).
+
+[6.2.1] - 2021-04-20
+====================
+- Fixed a minor bug in the linter.
+
+[6.2.0] - 2021-04-20
+====================
+- Fixed several glitches in paramspace implementation (handling of bools, returning scalar values) (@kpj).
+- Fixed bugs in module implementation (@dlaehnemann, @johanneskoester).
+- Fall back to greedy scheduling solver if ILP solver needs more than 10 sec (@johanneskoester).
+
+[6.1.1] - 2021-04-07
+====================
+- Fixed several small bugs of the new module system (@johanneskoester, @dlaehnemann).
+- Fixed archive based conda deployment (@johanneskoester).
+- Better handling of download and target attributed in the interactive report (@johanneskoester).
+
+[6.1.0] - 2021-04-01
+====================
+- Snakemake now uses **mamba** as the default conda frontend (which can be overwritten by specifying to use conda via the --conda-frontend flag) (@johanneskoester).
+- Profiles using --cluster option can now handle relative submit script paths in combination with arguments (@kdm9).
+- New AutoRemoteProvider, which infers the type of remote file protocol from the given URL (@kpj).
+- When using global container directive, container usage can be deactivated on a per rule base (@bilke).
+- Bugfixes for checkpoint handling (@johanneskoester).
+- Bugfixes for the module system (@johanneskoester, @dlaehnemann).
+- Various improvements for the tutorial.
+
 [6.0.5] - 2021-03-11
 ====================
 - Fix bug (introduced with 6.0) when handling of HTML directories in report (@johanneskoester).